AgenticAI/ingestion_service/requirements.txt
# ==========================================================
# FastAPI & Server
# ==========================================================
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9

# ==========================================================
# OpenAI
# ==========================================================
openai==1.30.1
langchain-openai==0.1.7

# ==========================================================
# LangChain Core
# ==========================================================
langchain==0.2.1
langchain-community==0.2.1

# ==========================================================
# Vector Database
# ==========================================================
faiss-cpu==1.7.4

# ==========================================================
# Document Loaders
# ==========================================================
pypdf==4.2.0
unstructured==0.13.6
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
pillow==10.3.0

# ==========================================================
# Database (SQL Server)
# ==========================================================
pyodbc==5.1.0

# ==========================================================
# Settings & Environment
# ==========================================================
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ==========================================================
# Utilities
# ==========================================================
typing-extensions==4.11.0

PyJWT==2.11.0